+++ example/README.md
@@ -1,0 +1,11 @@
+# CountUp and useCountUp example
+
+This example can be used for development or testing the component and hook.
+
+## Installation
+
+- Fork and clone the repo
+- Run `yarn`
+- Run `yarn start`
+- Visit http://localhost:8888/
+- Change the code in the `index.ts` file and _refresh the page_ to see your changes (There is no HMR).
