+++ example/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>use-elapsed-time dev server</title>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script src="js/index.js"></script>
+  </body>
+</html>
