--- src/components/index.ts
@@ -1,1 +1,0 @@
-export { CountUp } from './CountUp'
