--- src/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useCountUp } from './useCountUp'
-export { useFormattedValue } from './useFormattedValue'
