--- src/utils/constants.ts
@@ -1,2 +1,0 @@
-export const DEFAULT_DURATION = 2
-export const DEFAULT_START = 0
